fall back to size_t for strides

+++ src/gpuarray_blas_opencl_clblas.c
@@ -147,8 +147,8 @@ static int dgemmBatch(cb_order order, cb_transpose transA, cb_transpose transB,
 static int hgemvBatch(cb_order order, cb_transpose transA,
                       size_t M, size_t N, float alpha,
                       gpudata **A, size_t *offA, size_t lda,
-                      gpudata **x, size_t *offX, int incX,
-                      float beta, gpudata **y, size_t *offY, int incY,
+                      gpudata **x, size_t *offX, size_t incX,
+                      float beta, gpudata **y, size_t *offY, size_t incY,
                       size_t batchCount, int flags) {
   return GA_DEVSUP_ERROR;
 }
@@ -156,8 +156,8 @@ static int hgemvBatch(cb_order order, cb_transpose transA,
 static int sgemvBatch(cb_order order, cb_transpose transA,
                       size_t M, size_t N, float alpha,
                       gpudata **A, size_t *offA, size_t lda,
-                      gpudata **x, size_t *offX, int incX,
-                      float beta, gpudata **y, size_t *offY, int incY,
+                      gpudata **x, size_t *offX, size_t incX,
+                      float beta, gpudata **y, size_t *offY, size_t incY,
                       size_t batchCount, int flags) {
   return GA_DEVSUP_ERROR;
 }
@@ -165,31 +165,31 @@ static int sgemvBatch(cb_order order, cb_transpose transA,
 static int dgemvBatch(cb_order order, cb_transpose transA,
                       size_t M, size_t N, double alpha,
                       gpudata **A, size_t *offA, size_t lda,
-                      gpudata **x, size_t *offX, int incX,
-                      double beta, gpudata **y, size_t *offY, int incY,
+                      gpudata **x, size_t *offX, size_t incX,
+                      double beta, gpudata **y, size_t *offY, size_t incY,
                       size_t batchCount, int flags) {
   return GA_DEVSUP_ERROR;
 }
 
 static int hgerBatch(cb_order order, size_t M, size_t N, float alpha,
-                     gpudata **x, size_t *offX, int incX,
-                     gpudata **y, size_t *offY, int incY,
+                     gpudata **x, size_t *offX, size_t incX,
+                     gpudata **y, size_t *offY, size_t incY,
                      gpudata **A, size_t *offA, size_t lda,
                      size_t batchCount, int flags) {
   return GA_DEVSUP_ERROR;
 }
 
 static int sgerBatch(cb_order order, size_t M, size_t N, float alpha,
-                     gpudata **x, size_t *offX, int incX,
-                     gpudata **y, size_t *offY, int incY,
+                     gpudata **x, size_t *offX, size_t incX,
+                     gpudata **y, size_t *offY, size_t incY,
                      gpudata **A, size_t *offA, size_t lda,
                      size_t batchCount, int flags) {
   return GA_DEVSUP_ERROR;
 }
 
 static int dgerBatch(cb_order order, size_t M, size_t N, double alpha,
-                     gpudata **x, size_t *offX, int incX,
-                     gpudata **y, size_t *offY, int incY,
+                     gpudata **x, size_t *offX, size_t incX,
+                     gpudata **y, size_t *offY, size_t incY,
                      gpudata **A, size_t *offA, size_t lda,
                      size_t batchCount, int flags) {
   return GA_DEVSUP_ERROR;
@@ -197,16 +197,16 @@ static int dgerBatch(cb_order order, size_t M, size_t N, double alpha,
 
 static int hdot(
         size_t N,
-        gpudata *X, size_t offX, int incX,
-        gpudata *Y, size_t offY, int incY,
+        gpudata *X, size_t offX, size_t incX,
+        gpudata *Y, size_t offY, size_t incY,
         gpudata *Z, size_t offZ) {
     return GA_DEVSUP_ERROR;
 }
 
 static int sdot(
         size_t N,
-        gpudata *X, size_t offX, int incX,
-        gpudata *Y, size_t offY, int incY,
+        gpudata *X, size_t offX, size_t incX,
+        gpudata *Y, size_t offY, size_t incY,
         gpudata *Z, size_t offZ) {
   cl_ctx *ctx = X->ctx;
   clblasStatus err;
@@ -247,8 +247,8 @@ static int sdot(
 
 static int ddot(
         size_t N,
-        gpudata *X, size_t offX, int incX,
-        gpudata *Y, size_t offY, int incY,
+        gpudata *X, size_t offX, size_t incX,
+        gpudata *Y, size_t offY, size_t incY,
         gpudata *Z, size_t offZ) {
   cl_ctx *ctx = X->ctx;
   clblasStatus err;
@@ -288,15 +288,15 @@ static int ddot(
 
 static int hgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
                  float alpha, gpudata *A, size_t offA, size_t lda,
-                 gpudata *X, size_t offX, int incX, float beta,
-                 gpudata *Y, size_t offY, int incY) {
+                 gpudata *X, size_t offX, size_t incX, float beta,
+                 gpudata *Y, size_t offY, size_t incY) {
   return GA_DEVSUP_ERROR;
 }
 
 static int sgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
                  float alpha, gpudata *A, size_t offA, size_t lda,
-                 gpudata *X, size_t offX, int incX, float beta,
-                 gpudata *Y, size_t offY, int incY) {
+                 gpudata *X, size_t offX, size_t incX, float beta,
+                 gpudata *Y, size_t offY, size_t incY) {
   cl_ctx *ctx = A->ctx;
   clblasStatus err;
   cl_uint num_ev = 0;
@@ -325,8 +325,8 @@ static int sgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
 
 static int dgemv(cb_order order, cb_transpose transA, size_t M, size_t N,
                  double alpha, gpudata *A, size_t offA, size_t lda,
-                 gpudata *X, size_t offX, int incX, double beta,
-                 gpudata *Y, size_t offY, int incY) {
+                 gpudata *X, size_t offX, size_t incX, double beta,
+                 gpudata *Y, size_t offY, size_t incY) {
   cl_ctx *ctx = A->ctx;
   clblasStatus err;
   cl_uint num_ev = 0;
@@ -424,15 +424,15 @@ static int dgemm(cb_order order, cb_transpose transA, cb_transpose transB,
 }
 
 static int hger(cb_order order, size_t M, size_t N, float alpha,
-                gpudata *X, size_t offX, int incX,
-                gpudata *Y, size_t offY, int incY,
+                gpudata *X, size_t offX, size_t incX,
+                gpudata *Y, size_t offY, size_t incY,
                 gpudata *A, size_t offA, size_t lda) {
   return GA_DEVSUP_ERROR;
 }
 
 static int sger(cb_order order, size_t M, size_t N, float alpha,
-                gpudata *X, size_t offX, int incX,
-                gpudata *Y, size_t offY, int incY,
+                gpudata *X, size_t offX, size_t incX,
+                gpudata *Y, size_t offY, size_t incY,
                 gpudata *A, size_t offA, size_t lda) {
   cl_ctx *ctx = X->ctx;
   cl_event evl[3];
@@ -460,8 +460,8 @@ static int sger(cb_order order, size_t M, size_t N, float alpha,
 }
 
 static int dger(cb_order order, size_t M, size_t N, double alpha,
-                gpudata *X, size_t offX, int incX,
-                gpudata *Y, size_t offY, int incY,
+                gpudata *X, size_t offX, size_t incX,
+                gpudata *Y, size_t offY, size_t incY,
                 gpudata *A, size_t offA, size_t lda) {
   cl_ctx *ctx = X->ctx;
   cl_event evl[3];
